Codechange: Split palette handling to separate file.
