+++ src/CMakeLists.txt
@@ -325,6 +325,8 @@ add_files(
     order_gui.cpp
     order_type.h
     osk_gui.cpp
+    palette.cpp
+    palette_func.h
     pbs.cpp
     pbs.h
     progress.cpp
