+++ src/widgets/dropdown_type.h
@@ -13,6 +13,7 @@
 #include "../window_type.h"
 #include "../gfx_func.h"
 #include "../gfx_type.h"
+#include "../palette_func.h"
 #include "../string_func.h"
 #include "../strings_func.h"
 #include "../table/strings.h"
